converted CHatInput to component
